# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"GamepadDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"number": {
						"type": "string"
					},
					"assignedTo": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"number",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GamepadListResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/GamepadDto"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GamepadResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/GamepadDto"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ControllerActionResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PriceSegmentDto": {
				"properties": {
					"playerType": {
						"type": "string"
					},
					"startElapsedMs": {
						"type": "number",
						"format": "double"
					},
					"ratePerHour": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"playerType",
					"startElapsedMs",
					"ratePerHour"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TabItemDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"nameAr": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"qty": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"name",
					"price",
					"qty"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SessionDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"mode": {
						"type": "string",
						"enum": [
							"prepaid",
							"postpaid"
						]
					},
					"playerType": {
						"type": "string"
					},
					"startTime": {
						"type": "number",
						"format": "double"
					},
					"pausedAt": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"totalPausedMs": {
						"type": "number",
						"format": "double"
					},
					"targetDurationMin": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"priceSegments": {
						"items": {
							"$ref": "#/components/schemas/PriceSegmentDto"
						},
						"type": "array"
					},
					"tab": {
						"items": {
							"$ref": "#/components/schemas/TabItemDto"
						},
						"type": "array"
					}
				},
				"required": [
					"mode",
					"playerType",
					"startTime",
					"totalPausedMs",
					"priceSegments",
					"tab"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ConsoleDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"dailyTotal": {
						"type": "number",
						"format": "double"
					},
					"displayOrder": {
						"type": "number",
						"format": "double"
					},
					"session": {
						"allOf": [
							{
								"$ref": "#/components/schemas/SessionDto"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"id",
					"name",
					"type",
					"status",
					"dailyTotal"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ConsoleListResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/ConsoleDto"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ConsoleResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/ConsoleDto"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ConsoleActionResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UserPublicDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"role": {
						"type": "string"
					},
					"createdAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					}
				},
				"required": [
					"id",
					"username",
					"name",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginResponseDto": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"token": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/UserPublicDto"
					}
				},
				"required": [
					"success",
					"token",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginRequestDto": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserActionResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChangePasswordDto": {
				"properties": {
					"username": {
						"type": "string"
					},
					"currentPassword": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"newPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CurrentUserResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"user": {
						"$ref": "#/components/schemas/UserPublicDto"
					}
				},
				"required": [
					"success",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserListResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/UserPublicDto"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AccountDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"role": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "playstation-system-backend",
		"version": "1.0.0",
		"contact": {}
	},
	"paths": {
...
		"/api/v1/controllers": {
			"get": {
				"operationId": "GetAllControllers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GamepadListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Controllers"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "SaveController",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GamepadResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Controllers"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GamepadDto"
							}
						}
					}
				}
			}
		},
		"/api/v1/controllers/batch": {
			"post": {
				"operationId": "SaveAllControllers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GamepadListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Controllers"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/GamepadDto"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
...
		"/api/v1/controllers/{id}": {
			"delete": {
				"operationId": "DeleteController",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ControllerActionResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Controllers"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/v1/consoles": {
			"get": {
				"operationId": "GetAllConsoles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConsoleListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Consoles"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "SaveConsole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConsoleResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Consoles"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ConsoleDto"
							}
						}
					}
				}
			}
		},
...
		"/api/v1/consoles/{id}": {
			"get": {
				"operationId": "GetConsoleById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConsoleResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Consoles"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateConsole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConsoleResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Consoles"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ConsoleDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteConsole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConsoleActionResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Consoles"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/consoles/batch": {
			"post": {
				"operationId": "SaveAllConsoles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConsoleListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Consoles"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/ConsoleDto"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
...
		"/api/v1/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequestDto"
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/change-password": {
			"post": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserActionResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangePasswordDto"
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/me": {
			"get": {
				"operationId": "GetCurrentUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CurrentUserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/auth/accounts": {
			"get": {
				"operationId": "GetAccounts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "SaveAccounts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserActionResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/AccountDto"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}